mlx_accel_core: Added FPGA transfers and page boundary
A transfer may consist of more than one transaction.
A transfer starts as many transactions as it can when
it starts running, and if needed continues to start
more transactions as old ones are complete and their
TID is vacated.
A transfer's state is managed under a spin lock, so
its different transactions completion may happen in
interrupt context in different CPUs without problems

Signed-off-by: Ilan Tayari <ilant@mellanox.com>
Issue: None
Change-Id: I5db824ab98b8911f728921e3775ea5fff6bf8b19