mlx_accel_core: FPGA QPC now member of conn object; Added using reserved GIDs
The connection object now holds a copy of the FPGA's QPC struct.
This allows keeping and managing its contents from creation to
connection.
Also, initial PSNs are now synchronized between FPGA and CX4
Lastly, the sgid_index is now reserved explicitly in mlx5 driver,
so that ib_core is not aware of it, and it does not depend on
creation of a vlan netdev.

Signed-off-by: Ilan Tayari <ilant@mellanox.com>
Change-Id: Ia0c74c35d3cb5dd6ed9fa9ecf184a02c737e34c0
Issue: None